refactor: 로비 페이지 관련 구조 리팩토링
