+++ src/features/presence/types.ts
@@ -1,0 +1,9 @@
+export type OnlineUserStatus = 'WAITING' | 'PLAYING'
+
+export interface OnlineUser {
+  id: string
+  nickname: string
+  status: OnlineUserStatus
+  avatarText: string
+  avatarBackground: string
+}
