+++ src/pages/lobby/hooks.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query'
-import { getLobbyRooms, getLobbyUsers } from './api'
+import { getLobbyRooms } from './api'
 
 const LOBBY_QUERY_STALE_TIME_MS = 20_000
 
@@ -7,14 +7,6 @@ export function useLobbyRoomsQuery() {
   return useQuery({
     queryKey: ['lobby', 'rooms'],
     queryFn: getLobbyRooms,
-    staleTime: LOBBY_QUERY_STALE_TIME_MS,
-  })
-}
-
-export function useLobbyUsersQuery() {
-  return useQuery({
-    queryKey: ['lobby', 'users'],
-    queryFn: getLobbyUsers,
     staleTime: LOBBY_QUERY_STALE_TIME_MS,
   })
 }
