+++ src/pages/lobby/mockData.ts
@@ -1,4 +1,4 @@
-import type { LobbyRoom, LobbyUser } from './types'
+import type { LobbyRoom } from './types'
 
 export const mockLobbyRooms: LobbyRoom[] = [
   {
@@ -72,36 +72,5 @@ export const mockLobbyRooms: LobbyRoom[] = [
     currentPlayers: 3,
     maxPlayers: 4,
     isPrivate: false,
-  },
-]
-
-export const mockLobbyUsers: LobbyUser[] = [
-  {
-    id: 'user-1',
-    nickname: '마블왕',
-    status: 'WAITING',
-    avatarText: 'M',
-    avatarBackground: '#f6c8a9',
-  },
-  {
-    id: 'user-2',
-    nickname: '주사위마스터',
-    status: 'PLAYING',
-    avatarText: 'D',
-    avatarBackground: '#7f8ea3',
-  },
-  {
-    id: 'user-3',
-    nickname: '행운의여신',
-    status: 'WAITING',
-    avatarText: 'L',
-    avatarBackground: '#dbc4f8',
-  },
-  {
-    id: 'user-4',
-    nickname: '부동산왕',
-    status: 'PLAYING',
-    avatarText: 'R',
-    avatarBackground: '#8f7f77',
   },
 ]
