+++ src/pages/lobby/types.ts
@@ -1,7 +1,5 @@
 export type LobbyRoomStatus = 'WAITING' | 'PLAYING'
 
-export type LobbyUserStatus = 'WAITING' | 'PLAYING'
-
 export interface LobbyRoom {
   id: string
   title: string
@@ -9,12 +7,4 @@ export interface LobbyRoom {
   currentPlayers: number
   maxPlayers: number
   isPrivate: boolean
-}
-
-export interface LobbyUser {
-  id: string
-  nickname: string
-  status: LobbyUserStatus
-  avatarText: string
-  avatarBackground: string
 }
